OA scrappers update
Updated OgladajAnime scrapers: Created an full-featured scraper for series (can be easily applied to movies, which will be done next). Since partial scraper that was using tooltip data is deprecated now, it was replaced with full scraper in search function, missing only trailer url (which is rather insignificant, because first episode on full page is always the trailer). Content data is now stored in database and updated as content is scrapped, which cuts down data fetch time to nearly zero when fetching again the same content. Also applied some small fixes to multiple systems.