Fixed CDA player
Fixed CDA player, which (not long after last commit) started requiring Referer header in order to serve the video. Due to that, plus the fact that JS's fetch function treats this header as insecure (wonder why that might be) had to pass off downloading to server and then re-serve downloaded data to the client. Such a mess. Anyway, this is done, and server also saves the chunks to temporary database, so they could later be used in order to download the entire video (just need to remember to clear this table from time to time, or maybye move it to another .db file, or preferably both). Tested on 2 series, but it was suprising to me that no errors have surfaced during those test, so the code is probably still buggy... BUT IT'S TIME TO CELEBRATE. WOO-HOOO