Structural update
Applied large structural update of the database as well as some scrapper code. This was done in order to make retreiving content-parent related metadata quicker and make it less computation-heavy in exchange for a bit more data stored. All OA content is now properly normalized to regular classes (Movie, Series, Episode) instead of storing custom types (such as ONA and OVA), which makes it easier to retreive OA-based content data. Prepared the server architecture for keeping watch state of episodes.